+++ SampleApps/LJCFacilityManager/LJCFacilityManager/Forms/Person/Module/PersonModule.cs
@@ -88,7 +88,7 @@ namespace LJCFacilityManager
 		// Displays the ViewBuilder to edit the current view.
 		internal void DoViewEdit(ViewComboControl viewCombo, ViewInfo viewInfo)
 		{
-			int viewID = viewCombo.LJCSelectedItemID();
+			int viewID = (int)viewCombo.LJCSelectedItemID();
 			var viewEditor = new ViewEditorList(viewInfo.TableName, false)
 			{
 				StartupViewDataID = viewID
@@ -310,7 +310,7 @@ namespace LJCFacilityManager
 			PersonGrid.Columns.Clear();
 
 			// Get the view grid columns
-			mViewInfoPerson.DataID = ViewComboPerson.LJCSelectedItemID();
+			mViewInfoPerson.DataID = (int)ViewComboPerson.LJCSelectedItemID();
 			var gridColumns = mDataDbView.GetGridColumns(mViewInfoPerson.DataID);
 			if (null == gridColumns)
 			{
@@ -352,7 +352,7 @@ namespace LJCFacilityManager
 			RelationGrid.Columns.Clear();
 
 			// Get the view grid columns
-			mViewInfoRelation.DataID = ViewComboRelation.LJCSelectedItemID();
+			mViewInfoRelation.DataID = (int)ViewComboRelation.LJCSelectedItemID();
 			var gridColumns = mDataDbView.GetGridColumns(mViewInfoRelation.DataID);
 			if (null == gridColumns)
 			{
