+++ SampleApps/LJCUnitMeasure/LJCUnitMeasure/Forms/UnitSystem/UnitSystemComboCode.cs
@@ -93,7 +93,7 @@ namespace LJCUnitMeasure
 			if (UnitSystemCombo.SelectedIndex >= 0)
 			{
 				// Data from items.
-				int id = UnitSystemCombo.LJCSelectedItemID();
+				int id = (int)UnitSystemCombo.LJCSelectedItemID();
 
 				detail = new UnitSystemDetail(mManagers)
 				{
@@ -146,7 +146,7 @@ namespace LJCUnitMeasure
 					, MessageBoxIcon.Question) == DialogResult.Yes)
 				{
 					// Data from items.
-					int id = UnitSystemCombo.LJCSelectedItemID();
+					int id = (int)UnitSystemCombo.LJCSelectedItemID();
 
 					var keyColumns = new LJCDataColumns()
 					{
@@ -175,7 +175,7 @@ namespace LJCUnitMeasure
 			mParent.Cursor = Cursors.WaitCursor;
 			if (UnitSystemCombo.SelectedIndex >= 0)
 			{
-				id = UnitSystemCombo.LJCSelectedItemID();
+				id = (int)UnitSystemCombo.LJCSelectedItemID();
 			}
 			LoadCombo();
